Add should_skip_buildkite_pipeline_upload.sh
`should_skip_buildkite_pipeline_upload.sh` is a simple script that
checks whether given conditions are all true. This is useful in
determining whether we should run a buildkite pipeline or directly mark
it as passing. To use it in a pipeline, replace the buildkite steps
with:

```
agents:
  queue: new-central
  slurm_qos: "debug"
  slurm_time: "00:05:00"

steps:
  - label: "Skip CI If we are only modifying md or .github files"
    key: "skip_ci"
    command:
      - should_skip_buildkite_pipeline_upload.sh
    soft_fail: true

  - wait:
    continue_on_failure: true

  - label: 'Upload pipeline'
    command: |
      if [ $$(buildkite-agent step get "outcome" --step "skip_ci") == "passed" ]; then
        buildkite-agent pipeline upload .buildkite/pipeline.yml
      fi
```

Having a separate script will allow us to add more conditions in the
future and immediately propagate them to the repositories.